--- a/aten/src/ATen/cpu/vec/vec128/vec128_float_neon.h
+++ b/aten/src/ATen/cpu/vec/vec128/vec128_float_neon.h
@@ -420,16 +420,9 @@ inline Vectorized<float> Vectorized<float>::frac() const {
   return *this - this->trunc();
 }
 
-//Added sleef Implementation for Maximum
+template <>
 Vectorized<float> inline maximum(const Vectorized<float>& a, const Vectorized<float>& b)  {
-  if(!a.has_inf_nan() && !b.has_inf_nan()){
-    return USE_SLEEF(
-      Vectorized<float>(Sleef_fmaxf4(a, b)),
-      Vectorized<float>(vmaxq_f32(a,b)));
-  }
-  else{
-    return Vectorized<float>(vmaxq_f32(a, b));
-  }
+  return Vectorized<float>(vmaxq_f32(a, b));
 }
 
 // Implements the IEEE 754 201X `minimum` operation, which propagates NaN if